requirements.txt
dnspython
azure-mgmt-dns